Move CreatureDefines.cpp functions to Creature.cpp
