+++ src/game/CMakeLists.txt
@@ -168,7 +168,6 @@ set (game_SRCS
     Objects/Bag.cpp
     Objects/Corpse.cpp
     Objects/Creature.cpp
-    Objects/CreatureDefines.cpp
     Objects/DynamicObject.cpp
     Objects/GameObject.cpp
     Objects/Item.cpp
